requirements.txt
pytest
requests
pytest-html
pytest-xdist
jsonschema
pyyaml
